Fix undefined variable bug for checkbox variables
Issue: With WP_DEBUG set to true, it shows 'undefined variable' error notices for any checkbox that is unchecked.

Cause: default values are not set for the various '$tChecked' variables (no alternative to the if clause). So, the call to read those variables on admin.phtml will throw a PHP error for any that do not get declared.

Fix: Set those variables to an empty string ('') if they shouldn't be checked. (Using the ternary operator to keep the code minimal.)