Adds acceptance for already authenticated user
This adds code that honors an already successfull authentication.

As the first parameter to the login-function is either "NULL" when no
authentication attempt has happened or WP_Error when a previous
login-attempt has failed or WP_User when a previous login-attempt
already has been successfull. Therefore when the first parameter is an
instance of WP_USer we now instantly return that instance without
further processing.

This has been reported in issue #81 by
[litinoveweedle](https://github.com/litinoveweedle)