Add GroupSeparator option
The GroupSeparator value specifies the group separator used for the value in
'Groups' setting. It's default value is ',' (comma).

This value might need to be changed to something else like ';' when using attributes
whose values contain commas in them. For example, when using 'memberof' in GroupAttr,
the values are DNs which contain commas.