BugFix: Allow multiple usage of %s for username substitution
Currently the sprintf() will complain about too few arguments when multiple "%s" are used in the User login "Filter".
This happens when drafting a filter which would match upon uid AND mail - so that users could login via uid OR mail: `(|(uid=%s)(mail=%s))`. The solution is to use str_replace() to substitute all occurences.