Fix possible user-takeover
This fixes an issue where users created in wordPress could be taken over
by users with the same userid from an LDAP-Backend. This should in the
first case never have happened but in the rare cases where LDAP
authentication was introduced late it might be possible.

This checks whether the user that is about to log in is already marked
as an LDAP-User and if not (and when the flag to check this is set) the
authentication will be passed on to the next module.

This fixes #172