Add docker-compose setup
This adds the necessary files to be able to spin up a development
environment with a single "docker-compose up".

The environment contains a fully configured LDAP-Server. The
WordPress-Setup still needs to be automated.

Currently the WordPress-setup needs to be done manually but for a quick
and dirty solution this is already quite OK.