Allow Environment-Variables as LDAP-URI or part
This allows one to provide the LDAP-Uri or parts of it as
Environment-Variables.

If one wants to provide a complete LdapUri as environment-variable they
can use the pseudo-schema "env" followed by a colon and the name of the
envoronment variable like this: `env:LDAP_URI`.

If one wants to provide only parts of the LdapUri via environment
variables they can provide the appropiate part by inserting
`%env:[VariableName]%` in the LdapUri. So if you for example want to
provide the password via an environment variable `LDAP_PASSWORD` you would
use this LdapUri:

    ldaps://user:%env:LDAP_PASSWORD%@server/baseDn

Note that these inserted environment variales must not be URL-Encoded!
If you provide a complete LdapUri though, its parts *have*  to be
URL-encoded!