Adapt to PHP8.1
In PHP 8.1 the ldap-functions no longer return ressources but classes.
That is no issue within the API but becomes an issue when checking
whether something is a resource as that suddenly will not be one
anymore.

This commit replaces the checks for a ressource with the check whether
the property is NULL or not.