Use unfiltered roles-list
As the filter is applied to fetch only the roles that the current user
can assign to another user - which is not aplicable to what we are doing
here - the list can become empty which will result in wrong information
being assigned.

Therefore we now simply use the WP_Roles content to define the order of
the roles to then use the first role.