Remove ldap errors in behat tests
This change removes the ldap-errors that occured due to already existing
entries from previous calls. For that we first check whether the entry
already exists and if so, delete it so that it can then be recreated.

I opted for this approach instewad of keeping the old entries as the new
one might have slightly different information that would then break the
tests.