Fix problem with double search
Due to the change in authentication and authorization we need to search the
use4r a second time during the authorization phase. Should the user have been
identified with a different property than the one used for the WordPress
username, that would fail as we passed in only the WP_User that didn't have the
original parameter anymore.

I could instead also have created a filter myself based on the WP_User objects
username and the username attribute, now that i think about it. This seemed more
straight forward...